Basic NuSpec file for MvvmCross.
Installing this package in a project targeting MonoAndroid will not work until NuGet 2.3.  This also applies to a PCL when you have added the MonoAndroid XML file to its SupportedFrameworks directory.  A workaround for either of these issues is to install the package in a project type that is supported (Windows Phone or Windows 8 for example), and then to reference the DLLs in the solution's package folder from the MonoAndroid or PCL project.