Converted OnViewModelSet method from abstract to virtual.
With the new bindings it is somewhat inconvenient to have to override that method constantly.

When a view overrides the OnViewModelSet they should start calling the base implementation as well.
This is because in the future that will probably be in charge of notifying fragments when the data
is available.