Fixed FolderExists in FileStore for WindowsStore
The previous implementation tried to get parent folder and see if the
folder asked for exists within that... If the folder passed in as
FolderPath was the LocalStorage folder, then parent will become the root
which the application does not have access to and throw errors.

Another workaround would be to give the app access to the root folder,
but that seems to be not the best way to go about it.