Pass the java handle down to the base constructor to avoid a potentially nasty leak
I haven't seen this in the wild but it should prevent a disconnect between the java and c#
if instantiating the activity from java. I don't *think* we'd hit this unless starting an activity
from an intent on the java side of things and we happened to override a virtual method called
from one of the constructors.

See http://developer.xamarin.com/guides/android/under_the_hood/architecture/#Java_Activation