Adding Cirrious.MvvmCross.WindowsUWP.rd.xml in nuspec.
The UWP package requires the file "Cirrious.MvvmCross.WindowsUWP.rd.xml" be distributed with it in order for projects using it to build properly. When the project is open in Visual Studio, this file is found under the Properties node of Cirrious.MvvmCross.WIndowsUWP. It's copied to the output folder during Release builds.