Clear the DataContext when a ViewHolder is recycled
Without this change we would hold onto the old data context which would
leak the bound item.  In the case of a cleared ItemSource this would mean
holding onto the entirety of the old list as well.

Using the test case in #1343:

1. With this change the first time a Mono GC is fired after the itemsource
  has been cleared the Java Finalizer will run on the ViewHolder
2. The second Mono GC will collect the ViewHolder.

After 2 the original ItemSource will be collected.

I think this fixes the memory leak as reported in #1343 if an explicit
GC is run.