Clear bindings on ViewHolders when the RecyclerView is detached.
Keep WeakReferences of all ViewHolders created by the adapter.  When the
associated RecyclerView is detached from the window we clear the bindings.
Note that this does not Dispose of the ViewHolder since the Adapter doesn't
really manage the lifetime of the ViewHolder and there is no callback for
when the ViewHolder is destroyed.

Without this we leak target bindings.

Fixes https://github.com/MvvmCross/MvvmCross/issues/1379