Automatically hide menu
At the moment the menu stays open when you navigate away from it and the developer needs to specifically call close to hide it. With this commit the menu will automatically close unless the developer indicates (using the ShowPanel property on the MvxPanelAttribute class) that the menu should be kept open.

You can see this in action in the sample project. When navigating to the "Master view" the menu stays open. However when you navigate to the "Example view" the menu automatically closes.