Added UpdateChangelog task
You can run it with:

```
sh build.sh --target UpdateChangelog --since_tag=6.2.3 --github_token=abcd
```

If you have `CHANGELOG_GITHUB_TOKEN` set as environment, you can ommit the `github_token` argument.

`since_tag` is optional, but can be used to limit how much is added to the changelog.

If you are on a release branch it will set the future version to the version matching that release branch.