Enable the Visual Studio fast up-to-date check
These four projects condition their source files by target framework.

It used to be the case that Visual Studio's fast up-to-date check would only consider the source items and references of the _first_ target framework.

This is no longer the case. It was fixed in https://github.com/dotnet/project-system/pull/6862 which shipped in VS 16.10, and all target frameworks are now correctly handled.

So longer as your developers are using VS16.10 or later (including VS2022), you no longer need to pay the inner-loop performance penalty for these projects by scheduling them for build all the time.

I am actively working on improving the fast up-to-date check in Visual Studio at the moment, and would love to hear about any issues you encounter with it.